Provide fallback for weight dimension on older renderers
We don't need RequireApi(33) as this is remaining renderer feature that
we use in implementation that doesn't have fallback already.

Instead of using weights to represent % of the element's size, we can
use % of the screen size as a fallback, instead of not displaying
anything.

Additionally, add a fallback for TEXT_OVERFLOW_ELLIPSIZE.

Bug: 362969518
Bug: 391536651
Test: Screenshot + manually (attached in the bug)
Relnote: "Updated Material 3 components (graphicCard and avatarButton) to provide fallback when `weight` expand dimension is not supported (e.g. below API 33). Updated text component to fallback to TEXT_OVERFLOW_ELLIPSIZE_END when TEXT_OVERFLOW_ELLIPSIZE is not supported by the renderer."
Change-Id: I19e2cd328c6a6ff6022b51b9cab9347527f984ea