[Slider] Fixed thumb alignment to ticks
The current slider's thumb does not exactly falls on the ticks as the position is calculated relative to the entire width of the track and doesn't take into account the corners. As it is more visually pleasing to do so for the continuous slider, the fix is only applied when steps > 0 and properly calculates the position so it matches the ticks. Also there is no visual impact for the corners as the slider with steps snaps to the values and does not gradually show/hide the corners.

Fixes: 390993098
Test: screenshot tests

Change-Id: I52c14d3dccd34dcf3bfa6ff900668dd418646597