Replace swipeToDismiss parameter in HorizontalPager
Address API feedback.

- swipeToDismissEdgeZoneFraction parameter has been removed
- Introduce gestureInclusion parameter which takes in a
GestureInclusion object to determine whether the subsequent scroll
event should proceed or not (via the allowGesture function)
- New default behaviour for gestureInclusion / swipe to dismiss
defined within PagerDefaults.disableLeftEdgeOnFirstPage

This change also changes swipe to dismiss default functionality, this
is now only available on the leftmost edge of the first page in the
Pager, as per recent UX discussion. See b/390565236 for details.

Bug: 384911702
Bug: 390565236
Test: Modify existing tests

Relnote: "A breaking change introduced here is that the Pager component
is no longer full screen by default, instead you will have to add
Modifier.fillMaxSize() to achieve this behaviour. Another change is
that the swipeToDismissEdgeZoneFraction parameter on the
HorizontalPager has been removed. Instead a new parameter,
gestureInclusion has been introduced which can be used to achieve
necessary swipe to dismiss behaviour. gestureInclusion takes in an
instance of GestureInclusion which decides whether the subsequent
scroll event should proceed or not. Allowing a gesture means that the
Pager consumes it - otherwise, it will be handled elsewhere such as by
a swipe-to-dismiss handler. The new default behaviour is to disallow
gestures on the leftmost edge of the first page in the Pager, and hence
allow swipe to dismiss in this region. All other pages will permit the
Pager to consume all gestures, meaning Swipe to dismiss cannot be
triggered from them. To achieve Swipe to dismiss behaviour on all
pages, simply provide a custom GestureInclusion instance (see
PagerDefaults.disableLeftEdgeOnFirstPage) that ignores the current
page."

Change-Id: I9d3aae01b9115c898e6994da9143fb499953d2b8