Publish LazyLayoutCacheWindow APIs
Introduce LazyLayoutCacheWindow a way to control item caching in lazy lists. Introduced a new overload of LazyListState and rememberLazyListState which takes a LazyLayoutCacheWindow so users can configure the cache window for a given LazyList.

Relnote: Introduce LazyLayoutCacheWindow a way to control item caching in lazy lists. Introduced a new overload of LazyListState and rememberLazyListState which takes a LazyLayoutCacheWindow so users can configure the cache window for a given LazyList.

Test: Added tests for new API.
Change-Id: Ie2992a9098a7b1fcbd9a5d53ebececc5b06b7923