[GH] Don't drop retry/refresh events before uiReceiver is set
## Proposed Changes
Because uiReceiver is set asynchronously it's possible to race with a call to retry or refresh on the PagingDataPresenter.

Changed the implementation to record calling these methods and then forward the event when the uiReceiver is set.

## Testing

Test: ./gradlew test connectedCheck

## Issues Fixed

Fixes: b/330725007

This is an imported pull request from https://github.com/androidx/androidx/pull/754.

Resolves #754
Github-Pr-Head-Sha: b2603e5bfe3c4307f4d3141cfc84c3e37e5cb099
GitOrigin-RevId: 6cf65386377ffa13d3cff8ddd265af8685d38fa3
Change-Id: I46f7b0538108a070ca417f44113d723610b541ba