Use XrUiAutomation#injectMotionEventToWindow on Android XR
Android XR provides an platform extension @TestApi which acts as an
alternative to UiAutomation#injectInputEvent, and which pairs the
injected MotionEvent with the Accessibility window ID of the window
targeted by the event. This is necessary to reach XR's embedded
SurfaceControlViewHost windows, used for SysUI and more, with injected
touch events. See go/axr-fix-uiautomator for more details.

Fixes: 381945084
Test: ./gradlew :test:uiautomator:uiautomator:publish
Test: TAP Global Presubmit on cl/720792900
Test: Treehugger presubmit on ag/31501057
Test: atest runs of XR UIAutomator integration tests
Change-Id: I3762fda2f642df9433e66bfde86c267f3a492fa9