Add support to run on a StandardTestDispatcher
On an UnconfinedTestDispatcher, all coroutines that are started or
resumed on the test scheduler run immediately until their first/next
suspension point. On a StandardTestDispatcher, all coroutines that are
started or resumed are held on the test scheduler until the scheduler is
explicitly ran.

This requires a few calls to runCurrent() to meet expectations when
doing anything in a test that (could) involve(s) coroutine dispatching.
This should have no effect when running on an UnconfinedTestDispatcher,
because any task that would be run by runCurrent() on a
StandardTestDispatcher, would be executed immediately and not arrive on
the scheduler.

We also have to start the Recomposer UNDISPATCHED in the test setup,
because otherwise in most cases the Recomposer would start after
setContent is set. setContent performs an immediate initial composition
which now happens before the Recomposer is started. When the Recomposer
is then started, it will invalidate the world and recompose everything
again.

Bug: 254115946
Test: Updated CustomEffectContextTest and PhaseOrderingTest

Relnote: "Added `runCurrent()` to `MainTestClock` to run all due tasks
on the underlying scheduler. This is to support running tests on a
`StandardTestDispatcher`, where tasks are added to the scheduler rather
than executed immediately when they have a delay of 0ms. A test can be
setup to use a `StandardTestDispatcher` by
[creating](https://developer.android.com/reference/kotlin/androidx/compose/ui/test/junit4/package-summary#createComposeRule(kotlin.coroutines.CoroutineContext))
a `ComposeTestRule` and passing a `StandardTestDispatcher` to it. By
default, tests are run on an `UnconfinedTestDispatcher`, in which case
`runCurrent()` will never have to be called."

Relnote: "Added support for running tests on a StandardTestDispatcher.
Historically, the test framework was (and still is) set up using an
UnconfinedTestDispatcher, which differs from a production environment in
subtle ways. You can now change this setup by creating your own
StandardTestDispatcher and adding it to the `effectContext` when
[creating](https://developer.android.com/reference/kotlin/androidx/compose/ui/test/junit4/package-summary#createComposeRule(kotlin.coroutines.CoroutineContext))
your ComposeTestRule (or when calling
[runComposeUiTest](https://developer.android.com/reference/kotlin/androidx/compose/ui/test/package-summary#runComposeUiTest(kotlin.coroutines.CoroutineContext,kotlin.Function1))
)."

Change-Id: I334d0d149d263eb65c327f3dbba4fad154e8a559