PdfPageObject Scenario in PdfTestApp
As we close to completion for Pdf Page Objects and Annotation we
look ahead for testing. This CL is related to use of PdfPathObject as
a simple inkObject to validate the edit APIs in PdfRenderer. This
Scenario as of now validates add and get API in manual testing.
Going ahead plan is more extend this to all PageObjects, Annotations and more
parameter extensive testing. Crucial part of code have latest. Android SDK
dependencies hence is commented out.

Bug: 393082146
Test: Validated changes on Pixel 9 pro and Pixel Tablet
Change-Id: Ia09115d119ae47453ac0000bc494964c185982ae