Remove XAnnotationBox APIs
The XAnnotationBox APIs are convenient but lead to classpath confusion since it mixes processing classpath with compile classpath. In KSP2 it is specifically brittle since KSP annotation inspecting is by design dynamic as opposed to Javac's APIs that allows AST annotation analysis using actual annotation class via reflection but is warned against in Element.getAnnotation(Class<A>)'s documentation.

This change removes XAnnotationBox and its usages from Room, default annotation values are used explicitly in the codebase. This change does keep APIs that find an annotation via a KClass but they now return XAnnotation as these are generally safe and still convenient.

A significant benefit of this change is detection and validation of missing types in annotation values.

Bug: 329508813
Bug: 392989692
Test: Existing
Change-Id: I30ce845ea26f3666866e78f135e6b625de9c2c92