Add visibility platform binding
This allows developers to directly bind to the visiility status and use
it in expressions as dynamic data.

CL has diffs in other files as I run `javaFormater` on all files in
protolayout-expression.

Fixes: 393366572
Test: Added unit. Locally testing with implementation
Relnote: "We have added additional platform binding to receive the visibility status of full layout, whenever it's changed."

Change-Id: I250c33d50f6daf8a1ae12ca05acef3069ded68d9