Replace --api-version-names with --api-version-signature-pattern
Previously, `--api-version-signature-files` and `--api-version-names`
were required to be provided together as each item in the latter
provided the version name for the corresponding file in the format.
That complicated both this code and Metalava. This change switches to
use the `--api-version-signature-pattern` which gives Metalava the
information it needs to extract the version from the files, removing
the need for `--api-version-names` altogether.

Bug: 388441272
Test: Verified that the `apiLevels.json` files generated by the build
      were identical before and after this change.
Change-Id: Ic69263d4b1bf4689dba2cdf670b4d9e521917d96