Make sure token acquired during prewarm request processing is released
When we process a prewarm `RequestOpen`, we currently don't release the
token we acquired after completing the processing, even though the token
is never connected to a VirtualCamera. This can lead to accounting
errors in the WakeLock, and dangling cameras.

This CL addresses this, adds 2 new unit tests and fixes an existing
broken one.

Bug: 393711831
Test: Camera2DeviceManagerTest
Change-Id: Iac91f7e8eb1042f6c0d3c5d7bf532b6bc59d1541