Fix goAsync exception handling
The goAsync coroutine helper was not running the `block` with the
SupervisorJob parent context, but in a nested coroutine context created
by `launch`. This meant that exceptions still propagated up the job
hierarchy and triggered process crash.

This change updates the helper to run the block in a scope created by
`coroutineScope`, so that errors are rethrown in the current scope
instead of propagating to the parent scope of the `goAsync` call. This
also means that goAsync will properly wait for all jobs launched in
`block` to finish.

Bug: 393125740
Test: GlanceAppWidgetReceiverTest#errorInBroadcastReceiverDoesNotCrashProcess
Test: CoroutineBroadcastReceiverTest
Change-Id: Ibed1a7df75c8b2b8833b4bd8b8618e94b04d233f