Create Lint issue for calling Lifecycle::currentState in composition
This adds a Lint issue against calling `Lifecycle::currentState` in a
Composable, as it's almost always preferable to use
`currentStateAsValue().value` instead, which causes recompositions when
the Lifecycle state changes.

This is analogous to calling `StateFlow::value`, for which there is
already a Lint issue. This is mostly a copy-paste of that one.

Also adds a quick fix to StateFlowValueCalledInComposition.

Test: Detector has test, based on ComposableStateFlowValueDetectorTest

RELNOTE: N/A

Change-Id: Iad484fad72149f67f263fc3e0d4799e939e93f95