Dispatch MotionEvents after async transfer from client in Z-Below mode
In Z-Below mode, the MotionEvent objects are being transferred from the
client app to the sandbox process to be dispatched on the SDK View.

The events are transferred after being received by the SurfaceView to
avoid the need for any transformation.

The events are targeting to be dispatched on the following frame (after
the one they are originally targeting in the client app), that to
improve the smoothness of the dispatching (number of processed events
per frame).

This change introduces a new internal class ProviderViewWrapper, which
replaces TouchFocusTransferringView, the current role of this new
wrapper is to dispatch the transferred MotionEvents in case in Z-Below,
while it does not do anything in Z-Above (Meaning that, the gesture is
no longer transferred from provider to client in case of Z-Above, making
client apps not scrollable in Z-Above mode, that should be fixed by
b/374270009).

Relnote: "UI providers start to get user interaction MotionEvent objects
in Z-Below mode. Client UI is not vertically scrollable in Z-Above mode."

Bug: 394321328
Bug: 345688233
Test: new Unit tests ProviderViewWrapperTest and SandboxedSdkSurfaceView
Change-Id: If3cd6f4da6007e3acf47bf44fc3b00412e5abb6b