Fix issue with sticking fling in Scrollable.
A child scrollable cannot stop the outer scrollable from flinging, that is, if the outer scrollable is flinging and the inner scrollable sends an onPostFling event, the outer scrollable won't consume it as we understand it is already consuming a velocity during animation. This is an existing behavior. This change fixes a bug where the outer scrollable might get stuck "thinking" that it's flinging, therefore indefinetely disallowing the inner scrollable from propagating its fling velocity during onPostFling.

Test: Added new test.

Fixes: 394307899
Change-Id: Ib75925a2bee5dc63f1d1af9d370bc7ca942e5362