[1/n] Add SharedUiAdapter annotated interface generation to tools-apicompiler
Adds code generation for SDK-defined interfaces annotated with
@PrivacySandboxInterface that implement the SharedUiAdapter interface.
Similarly to SandboxedUiAdapter, special support for it is needed to
hide to/fromCoreLibInfo calls from both UI client and provider.

This change covers StubDelegate and CoreLibInfoAndBinderWrapperConverter
generation, which is validated in tools-apicompiler tests. Aidl
generation, as well as usage of the interface as a method
parameter/return type will be covered in the following changes.

To allow code generation, a new type, Models.sharedUiAdapter, is added.
sharedUiAdapter is also added to the list of valid supertypes.

Bug: 365553832
Test: ./gradlew :privacysandbox:tools:tools-{apigenerator,apicompiler,apipackager,core}:test

Change-Id: I2309634c7b7c9284078f5872d13db109044057bb