Replace pagerSnapDistance with maxFlingPages parameter
Address API feedback.

This simplifies the usage in the Wear Pager, and only the maxFlingPages
Int value needs to be kept in memory. In snapFlingBehaviour a
Compose PagerSnapDistance object is instantiated from this. If in the
future the Wear pager needs more complicated PagerSnapDistance usage, a
Wear PagerSnapDistance should be introduced.

Bug: 384911702
Test: N/A

Relnote: "When using PagerDefaults#snapFlingBehaviour the
pagerSnapDistance parameter has been replaced with an Int parameter
maxFlingPages, which can be used to specify the max number of pages the
Pager should fling."

Change-Id: I8cfc03982b9f75488c588303bf8da0e160af623b