Check for error types along with none when searching supers.
When inspecting super types for indices and primary keys, Room should abandon searching when the super class is an error type, which can happen on a round that has not yet generated the super. This avoids further errors and eventually lets Room deffer generation to another round.

An actual case of this can occur with Kotlin's serialization library.

Bug: 365423060
Test: QueryTest in Room's KMP test app
Change-Id: I07fb39a8923eee5dde92736364d72f55fe0c10c0