Fix: Getting DeadObjectException if the service is disconnected and we try to scroll through the document
Page should be released from BitmapFetcher only if the service is
connected. If the service would have been disconnected or killed, the
pages would already would have been released and a new service will be
launched. If the connected is broken then the pages would be released when the connection is re-established
So this invocation is only required if the service is connected already.
This is blocking call so if the page is being released while the service
is disconnected, it leads to the UI thread being blocked indefinitely.

Bug: 394312168
Test: Manual only and cant kill bound service from an automation
Change-Id: Iba4ecc1b7d5a1a21efd70563d109631ae4c04812