Fix detection of @Relation with a Set
The assignability check in the relation collector was wrong and since in aosp/3464673 it type was changed to MutableSet it would fail the check as the order matters, mutable collection can be assigned to non-mutable but not the other way around.

Also fix empty collection creation, the collector should compare raw types and not full field type name as that includes type params.

Bug: 394885249
Test: DaoRelationshipKotlinCodeGenTest.kt
Change-Id: I0ccbcf76c3100098785acfe99d14a8279ab663b7