Padding fix for FloatingToolbars
The fix ensures that the paddings are visually balanced when collapsing
a Floating toolbar.
A new internal modifier was added to ensure padding are added and
animated to the main `content` when the `leadingContent` or
`trailingContent` is missing or hidden.

Also, switched to use `heightIn` and `widthIn` instead of `height` and
`width` when setting the toolbar's height or width. This allows a more
flexible component that can grow in case its holding a larger
icons/components.

Internal changes:
1. AlignmentLines were added to the minimumInteractiveComponentSize in
order to allow us reading them and applying additional padding when
needed.
2. A minimumInteractiveBalancedPadding modifier was added to the
   FloatingToolbar, and we use it on the toolbar's content to ensure
   that when the content is collapsed the paddings are balanced.

Fixes: 394930575
Test: updated
Relnote: "Fixed Floating Toolbar padding to ensure visual balance during
collapse, and improved flexibility for larger content."

Change-Id: I06c00c626b67d7eaa98fd4969806ad7b00862b74