Add fallback for built-in types using SerializersModule and format-specific types handling
Use pre-registered contextual and polymorphic serializers to support these built-in types:

On common:

- SavedState

On Android:

- Size
- SizeF
- CharSequence (interface)
- java.io.Serializable (interface)
- Parcelable (interface)
- IBinder (interface)
- SparseArray<Parcelable>

After this users don't need to specify serializers when encoding/decoding these types at root. Or when encoding/decoding interface types anywhere.

Note that when declaring a serializable object with properties of non-interface built-in types a serializer is still needed. Users can use our built-in serializers in this case. For example:

```
@Serializable data class Foo(@Serializable(with = SizeSerializer::class) val bar: Size)
```

Or just use contextual serializer:

```
@Serializable data class Foo(@Contextual val bar: Size)
```

Also note that `MutableState` and `MutableStateFlow` support is not included in this CL. We may consider adding that in the future.

Relnote: Add fallback for built-in types using SerializersModule
Test: SavedStateCodecAndroidTest
Change-Id: Ic01d267d1f3a18e7a23ec0866076bbd2c940e5e6