Add Tablet portrait in Preview Screen sizes
The multi preview annotation class PreviewScreenSizes is updated with
the addition of the portrait tablet form factor.
The default Tablet was in Landscape mode. To keep name consistency
with the other form factors such as "Phone" and "Phone - Landscape",
"Tablet" is now renamed in "Tablet - Landscape" while the new added
portrait mode takes the default name "Tablet".

Bug: 381123648
Test: existing
Relnote: "Add Tablet Portrait in the collection of Preview Screen sizes to help the developer to consider bigger windows"

Change-Id: Ia1976dfd73960d190004e59269e042ab5fbc457c