Fix bug in wrapper provider's enabled state
After launching the sample app, wrapper routes would appear in the
output switcher, but the toggle would show the service as disabled. The
workaround was to toggle the enabled value twice (false -> true ->
false). But the issue would return sporadically, presumably as a result
of shared preferences being cleared.

Taking a step back, shared preference fights the package manager as the
source of truth for whether the service is enabled or not, so
ultimately, it's best to use package manager as the source of truth for
that.

This change:
- Uses package manager as the source of truth for the
  provider-enabled-toggles' values.
- Is a partial revert of aosp/3349767.
- Simplifies the SettingsActivity by deduplicating code.
- Calls setMediaTransferRestrictedToSelfProviders whenever the wrapper
  provider service enabled state changes.
- Makes the wrapper provider be disabled by default (I believe that was
  the original intent).

Bug: 205124386
Test: manually using the app
Change-Id: Ie299652aa28bc411315472634a29d2140130349e