Address test exception forwarding feedback
This CL addresses both g3 and API council feedback on the test exception
changes made in aosp/3106918.

For g3, we will no longer wrap the forwarded exception, and prefer the
original exception instead. Follow-up exceptions are still included as
suppressed exceptions, but with an added separator to indicate that
they may be a cascading exception.

For API Council, the `UncaughtExceptionHandler` is no longer a `fun`
interface. Additionally, because of the change for g3, the
composition phase enumeration has been removed as it was no longer used.

Fixes: b/395697927
Test: UnhandledExceptionTest, JUnitUnhandledExceptionTest
Relnote: N/A
Change-Id: I1efd1e3bc6187f3ba6fbc5d84187372c3debfcec