Rename SnapshotObserver methods to match API guidelines
Test: Refactor, existing tests pass
Fixes: b/373931333
Relnote: """Renamed SnapshotObserver methods onCreating to onPreCreate
and onDisposing to onPreDispose to match API guidelines. The previous
methods are deprecated and the new methods default to call the old ones
so exisiting implementations of this interface will continue to work
until the methods are removed.

These methods will be removed in 1.9.0 alpha in a follow-up CL that will
shortly follow this one. They  will be removed from 1.8 beta before 1.8
stable."""

Change-Id: I6d753f73ac673337dcf5d35aa688b9dfc13ef24a