Early terminate animating scroll delta when possible
When back scroll is necessary in lookahead, we create
an animation to gradually apply the scroll delta to
approach pass to avoid jarring visual jumps.

However, when the lookahead pass and approach pass have
the same measurements for the items, such scroll delta
animation will have no effect. This change specifically
checks for cases where the scroll delta animation
between two passes is redundant and early terminates
the animation.

Test: New test added
Bug: 396169351

Change-Id: I6fa3afde8ef6fff0727246b9f1bf2af29d00a447