Resolve video and audio capture/encode frame/sample rates
This commit introduces the resolution of video frame rates and audio sample rates into separate capture and encode values.

[Video]

The inputs for video frame rate are:

* Capture frame rate: Derived from `StreamSpec.getExpectedFrameRateRange`, originating from the "setTargetFrameRate" API.

* Encode frame rate: Derived from `VideoSpec.Builder.setFrameRate(int)`. We will later have a public API to set this value. If not set, it defaults to the capture frame rate.

The capture/encode ratio, calculated from these frame rates, represents the video speed adjustment.

Note: This change modifies the original video frame rate resolution logic when CamcorderProfile is found. Previously, it adopted the CamcorderProfile's frame rate. This change bases on the `StreamSpec.getExpectedFrameRateRange`'s upper bound if the range is set. This is because the expected frame range's upper bound represents the actual frame rate output from camera, but it may not the profile's frame rate.

[Audio]

Audio sample rate resolution refers to video's capture/encode ratio. The intended encode sample rate is sourced from `AudioSpec.getSampleRate` or defaults to CamcorderProfile's sample rate.

Although an API for setting `AudioSpec` sample rate is not in the scope, this change accounts for it in the capture and encode sample rate calculations.

Bug: 400369088
Test: ./gradlew camera:camera-video:connectedAndroidTest
Change-Id: I4b55f7ac53b099d343dacf74ebb3d4d398341b32