Make the rest SavedStateReader.getXXX() functions throw when value type doesn't match return type
We throw IllegalStateException when the saved value's type does't match function's return type in most of our `SavedStateReader.getXXX()` functions, but return a pre-defined default value in the following functions on Android:

- getBoolean
- getChar
- getDouble
- getFloat
- getInt
- getLong

In this CL we make these throw as well so the behaviors for all `getXXX()` are consistent. If a default value is preferred users should use the `getXXXOrElse` variants.

To achieve this on Android without using deprecated APIs we call the corresponding functions in `Bundle` twice when it's possible that the value does not have a correct type. With some optimization it's actually faster than the current version in most cases.

Here's my test on a Pixel Fold using Microbenchmark:

              current    this CL(no collision)   this CL(collision)
getBoolean    43.1 ns                 22.7 ns               44.5 ns
getChar       44.8 ns                 23.6 ns               46.0 ns
getDouble     42.8 ns                 22.7 ns               45.8 ns
getFloat      43.1 ns                 23.0 ns               46.0 ns
getInt        43.4 ns                 22.9 ns               44.4 ns
getLong       43.9 ns                 23.9 ns               45.2 ns

"Collision" here means the saved value is the same as our chosen default value (when we need to call `getXXX` twice). This should be like 50% of time for Boolean and less likely for other types. There is a minor slowdown (~3% to 7%) when collision happens, but ~45% to 47% improvement otherwise.

Test: SavedStateTest
Bug: 399317598
Relnote: Make the rest SavedStateReader.getXXX() functions throw when value type doesn't match return type
Change-Id: I78c4adedf59bc93ae8851ef105b7ef362bd37a64