Move tracing-driver's WireTraceSink to separate module
Test: Tests in tracing-driver-wire
Relnote: N/A

This allows the tracing interface to be used (e.g. across various KMP
libraries that would want KMP tracing) without taking a dependency on
wire. This will enable apps that want to use com.google.protobuf or
some entirely different and custom serialization to avoid this
dependency.

Change-Id: Id9669a87b72a86e5349573acedeae71d8c3a14e6