Share library versions between ui-client and ui-provider
Any version of ui-client can establish session with
any version of ui-provider.
To allow session to be established succesfully ui-client need
to know which version of ui-provider it is interacting and vice-versa.
This change allows ui-client to share it's version to ui-provider and
ui-provider shares it's version to ui-client when session is being created.

Test: Unit tests
Bug: 390102298
Bug: 375389719
RelNote: NA

Change-Id: I78b05821d6e193c20754fb5a0583019b43544b05