TraceValue API
Introduce trace value API for marking/logging trace with values at a given point in time. Trace values show in Prefetto as a new lane that can be seen through the timeline of a trace. On Android they're implemented using Trace.setCounter.

Test: N/A
Relnote: Introduce trace value API for marking/logging trace with values at a given point in time. Trace values show in Prefetto as a new lane that can be seen through the timeline of a trace. On Android they're implemented using Trace.setCounter.

Change-Id: Idcf48360d3bc608f6397a5ba7368ba7246c3712e