Add basic integration tests and fix found issues.
Found issues:
1. wrong xml file name
2. AppFunction annotation should be AnnotationRetention.BINARY
3. A bug that we don't support void return type

Impl notes:

There are two packages: androidx.appfunctions.integration.testapp
(the package to be tested) and
androidx.appfunctions.integration.testapp.test (the test package).
he instrumented test is running as the process of the package to be tested, instead of that of the test package.
Regardless of whether we use targetContext or context to make the app
function call, the system server will see the calling UID as the target
app's UID. In fact, using context to make the app function call would
crash because the system server finds that the package name does not
belong to the calling UID.
Thus, this change has the package to be tested execute functions
from its own process.

Bug: 397634300
Test: Run the newly added tests.
Relnote: N/A

Change-Id: Ib2130c7f86f8e2a0abd75a17071000b412e6daf6