Set up providerevents and providerevents-plau-services modules
providerevents - module that holds an abstract class that must be
extended by credential providers to receive requests.
providerevents-play-service - module that contains a class that provides
an interface that can be used to call APIs on the credential provider's
service after binding to it. This module is play-services specific and
is used when play-services is fulfilling credential manager requests.

Test: local testing
Relnote: "Add new providerevents module for credential providers to
extend a service and receive updates on credentials."

Change-Id: I8cd3a1a780ee1eb177c031340a169a6a105cebd6