Disable Swipe to dismiss when talkback enabled on Pager
- Copy TouchExplorationStateProvider into foundation to get support for
checking whether TouchServices are enabled in foundation

Bug: 384911936
Test: N/A - Added in next CL in chain

Relnote: "The new behaviour for the Pager is to ignore gestures on the
left edge of the first page, this allows swipe to dismiss handlers to
take over if present. However if talkback is enabled then the Pager
will always consume gestures, never allowing swipe to dismiss handlers
to take over. PagerDefaults.ignoreLeftEdgeOnFirstPage has been renamed
to PagerDefaults.gestureInclusion."

Change-Id: Iee486edeae3111a069959bc5f516ff20f4652747