Add support for session observer signal options
Adds the ability for an owner of a SessionObserverFactory to specify the
signals that they wish to receive events for. The set of available
signals is defined in ui-core as a Set of Strings. These options are set
on the SessionObserverFactory that is registered using
addObserverFactory.

After the UI session is active in the client container, the client
container will send a set of strings to each SessionObserver, using the
SessionObserverContext object. These strings represent the signal
options which are supported by the client container.

This change also delays the sending of SessionObserver.onSessionOpened
until after the first frame has been rendered after the session has
opened on the client container.

Also adds logic to AbstractSandboxedUiAdapter to delegate callbacks to
each SessionObserverFactory's observers based on their signal options.

Relnote: "Added support for session observer signal options."
Test: Manual, SandboxedSdk(View/Ui)Test, IntegrationTests
Bug: 341895747
Change-Id: I53cb787292de3376e26df476b80d21d64ab7df15