Add CurvedModifier.clearAndSetSemantics for use with Material3 timeTextCurvedText and timeTextSeparator
Unlike CurvedText, TimeText should not interact with TalkBack. Updated timeTextCurvedText and timeTextSeparator so that semantics are not automatically added to the CurvedSemanticsProperties.

BUG: 399325647
Test: Added samples to integration demos and tested manually
Relnote: "We have added CurvedModifier.clearAndSetSemantics to provide a means by which curved semantics can be turned off. CurvedText continues to default the content description to the text, but timeTextCurvedText and timeTextSeparator do not now announce their contents."
Change-Id: I4b568e3cced9e028b6efefce3c9f6150172ee05b