Updates logic for multiple transfer of uiAdapters
Today, UIAdapter transfer to a process and back leads to chained sessions.

This change ensures that if a ClientWrapperAdapter is sent back to the process
from where it originated, it wouldnt need another binder and the old one
can be reused.

This change introduces an interface CLientWrapperAdapter implemented by
both Local and Remote client side uiAdapters. The change checks
if the uiadapter being sent in a ClientWrapperAdapter and if so,
it re-uses the bundle that was used to create that client uiAdapter.
This ensures there is only one binder object ever created for a
uiadapter.

We will further need to add support for uiAdapters base class a
return/param type b/397610151 to ensure these can be forwarded smoothly
for mediation and other usecases

Bug: b/398865687
Test: Tested locally by sending a uiadapter back from a remote process
      and ensured it uses localadapter to render.
Change-Id: I68ca4cebf2b217f9a16bd2787b05dfb53eed5e1a