Fix TextAutoSize Ellipsize Bug
We previously performed auto size layout with TextOverflow.Clip and only used the user-supplied overflow for the final pass. We actually need to consider the overflow mode during auto size layout to decide whether it did overflow or not, so we now use the MultiParagraphLayoutCache's `overflow` property for auto size layout too.

Additionally, fixed the logic for StartEllipsis and MiddleEllipsis in single-line vs multiline text.

Test: New in MultiParagraphLayoutCacheTest
Relnote: Fixed a bug in BasicText and TextAutoSize where using TextOverflow.Ellipsize would result in smaller-than-optimal text.
Fixes: 396582066
Change-Id: I1e1d8710f8719edb229d4826da746d195875590b