Address API council feedback : Add `requireOffset` to SwipeToDismissBoxState
`requireOffset` method was added to SwipeToDismissState as a recommended way of getting an offset of SwipeToDismissBoxState.
An `offset` field was left in the api for cases when a value with NaN is required

Bug: 389659374
Test: None
Relnote: "Add `requireOffset` to SwipeToDismissBoxState. `requireOffset` method was added to SwipeToDismissState as a recommended way of getting an offset of SwipeToDismissBoxState"
Change-Id: I21042e15e23edad7cfed53210d4768b1d75d4a6c