Clarify KDocs of iconButton component on defaulting to dynamic colors theme
This only updates iconButton component to align on the wording of this
clarification. Developers should know that components' default is
dynamic theme colors, with a fallback option when theme is not
available.

Once this is approved, I will update all KDocs of components.

Bug: 353703413
Bug: 350034071
Test: KDocs changes only
Relnote: "Clarification of default dynamic color theme in ProtoLayout Material3 components."
Change-Id: I9d831a7a73a5f6f06af872a27ca226f3d7365583