Fix testCircularScrollingGesture on medium emulators.
I'm assuming that the gesture nav insets have changed slightly, so the
circular scrolling gesture doesn't work properly. Changing the gesture
to swipe across the top of the view (which should not be covered by any
insets) rather than down the side fixes the issue, and keeps the spirit
of the test.

Bug: 399667952
Test: Run androidx.wear tests.
Change-Id: Ifc1fc59bd7bb090de3a0b4854bac2edfc97330ba