Add generics to EntryProvider
The entryProvider DSL currently erases the type of the NavEntry returned
by the EntryProviderBuilder. This makes it incompatible with 1)
SinglePaneNavDisplay with a specified type or 2) a lambda that returns a
NavEntry with a specified type.

We should add generics to the entryProvider and the EntryProviderBuilder
to ensure that they can both be used with specific types.

This also allows us to change the entryProvider parameter in
SinglePaneNavDisplay to return a NavEntry with just the type instead of
out type.

RelNote: N/A
Test: Updated test that use entryProvider to have a type
Bug: 397911582
Change-Id: I864eee51783234ebedb02d0af0c38741d424befd