Fix deeplink matching logic
Right now deeplink does not match if the request has non-null fields that are null in the deeplink. However this should still match.

A Deeplink should be a potential match as long as any added non-null fields (uri / action/ mimeType) are also non-null and matches in the request.

Reverted a few test updates from aosp/3490354 and reorganized deeplink-related tests to improve readability and prevent false positives.

Test: ./gradlew navigation:navigation-common:cC
Test: ./gradlew navigation:navigation-runtime:cC
Bug: 399826566
Fixes: 399136426
Relnote: "Fixed deeplink not matching when the request has non-null fields that are null in the expected deeplink."
Change-Id: Ie5e36911112218baf8c53ee142f1366caa31a8b2