Fix KSTypeVarianceResolver for typealias with @JvmSuppressWildcards.
In KSP1, @JvmSuppressWildcards annotation was copied from the type being
aliased to the typealias itself. However, in KSP2, that does not happen.
The KSP2 behavior actually seems more correct, so rather than file a
KSP bug, I think we should fix this in XProcessing. In particular,
XProcessing shouldn't rely on the incorrect KSP1 behavior to work, so I've
updated the code to keep track of the annotations manually to make things
work with KSP2.

I've also removed a workaround for https://github.com/google/ksp/issues/1376
which is has been fixed in KSP for a while now.

Bug: 394692559
Test: KSTypeNamesGoldenTest.kt
Change-Id: Id1abbbac330c334a2ee6493d6e78e609e0c0c404