Move all built-in List and Array serializers out of public API
Lists and Arrays are serializable by default and users shouldn't need to explicitly specify serializers for them. So we're changing the visibility of these built-in serializers from public to internal:

* ParcelableArraySerializer
* ParcelableListSerializer
* CharSequenceArraySerializer
* CharSequenceListSerializer

The instances of corresponding types will still be optimized while serializing, no matter they're at root level or as a property of another serializable class.

Test: existing tests still pass
Relnote: Move all built-in List and Array serializers out of public API
Change-Id: Ida2938fa2ad727ddfc985ff11d13b8840a383618