Remove tabularNum as default font setting in Typography.NUMERAL_*
UX spec is that default should be non tabular/non monospace to
take advantage of more space. The only cases when tabular should
be used is when the text is animating.

See b/393024554 for context. This is already done in Compose and
talked with UX on making sure we need the same for Tiles and
known implications (i.e. developers need to set it manually).

Fixes: 400434540
Test: Screenshots
Relnote: "Typography.NUMERAL_* typographies are no longer tabular/monospace by default. If text is animating, it is highly recommended to add `FontSetting.tabularNum()` setting to it. In all other cases this monospace option is not needed and there will be more available characters by not using it."
Change-Id: Id3cd9c16f8095200e396152e5df7475c3b93df8f