Move root lint.xml config to buildSrc subdirectory
Having the file directly in buildSrc causes issues for `./gradlew buildSrc-tests:lint` when lint checks are set to run on tests in lint.xml. The configuration file was registered twice, once from the setting in LintConfiguration and once from scanning the directory. When it was picked up from scanning the directory, it caused the lint warning "The tests attribute can only be specified for lint.xml files at the module level or higher".

Bug: 382507748
Test: `./gradlew buildSrc-tests:lint` for topic
Change-Id: I0629ac180487ac200b263caf3030739ee71c292e