Update margins in M3 primaryLayout to be calculated as ceil
Margins should be rounded up when calculated as percentage of the
screen size to be generous enough to avoid clipping. This is a UX
requirement.

This can have up to 1dp each side affect on tiles UI.

Bug: 376713820
Test: Screenshot
Relnote: "Margins for primaryLayout are now properly rounded up instead, which can have effect on some layout up to 2dp space less for the main slot."
Change-Id: I8f5d39e5a019e783a40fa5392fd4686c0d1b82ea