Add logic to measure obstructions on UI containers
Added a signal option to support the calculation of obstructions on
SandboxedSdkViews. Obstructions are rectangles of a non-zero size and
opacity that occlude the visibility of the UI container. No occlusions
will be reported if the view is in Z-above mode. Also adds a
corresponding obstructedGeometry field to SandboxedSdkViewUiInfo.

When registering a view with the geometry measurer, the client should
specify whether or not obstructions should be calculated for this view.
The geometry measurer will perform a tree walk if any client requires
obstructions data. The geometry measurer will perform selective pruning
for performance reasons, such as not walking a subtree if it does
not intersect with any target view, or if it is fully transparent.

Also added a hidden fragment to the manual test app for testing.

Relnote: "Added logic to measure obstructions on UI containers."
Test: Manual, SandboxedSdkViewTest, SandboxedSdkViewUiInfoTest
Bug: 341607865
Change-Id: I34bea6c7c016b1a7a03056ce7e8256e644235b74